move project inner to separate component
